+++ gotatun/src/device/tests/mock.rs
@@ -28,6 +28,8 @@ use tokio_stream::wrappers::BroadcastStream;
 use x25519_dalek::{PublicKey, StaticSecret};
 use zerocopy::IntoBytes;
 
+use crate::key::PeerPublicKey;
+
 use crate::{
     device::{Device, DeviceBuilder, Peer},
     noise::index_table::IndexTable,
@@ -130,11 +132,11 @@ pub async fn device_pair() -> (MockDevice, MockDevice, MockEavesdropper) {
     let pubkey_a = PublicKey::from(&privkey_a);
     let pubkey_b = PublicKey::from(&privkey_b);
 
-    let peer_a = Peer::new(pubkey_a)
+    let peer_a = Peer::new(PeerPublicKey::new(pubkey_a).unwrap())
         .with_endpoint((endpoint_a, port).into())
         .with_allowed_ip(Ipv4Network::new(Ipv4Addr::UNSPECIFIED, 0).unwrap().into());
 
-    let peer_b = Peer::new(pubkey_b)
+    let peer_b = Peer::new(PeerPublicKey::new(pubkey_b).unwrap())
         .with_endpoint((endpoint_b, port).into())
         .with_allowed_ip(Ipv4Network::new(Ipv4Addr::UNSPECIFIED, 0).unwrap().into());